2 commit

+++ src/main/java/ru/vanek/task_management_application/services/impl/CommentsServiceImpl.java
@@ -1,6 +1,9 @@
 package ru.vanek.task_management_application.services.impl;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.prepost.PreFilter;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import ru.vanek.task_management_application.dtos.requests.CommentRequest;
 import ru.vanek.task_management_application.dtos.responses.CommentResponse;
 import ru.vanek.task_management_application.models.Comment;
@@ -12,7 +15,9 @@ import ru.vanek.task_management_application.utils.CommentConverter;
 
 import java.util.Date;
 import java.util.List;
-
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class CommentsServiceImpl implements CommentsService {
 
 
@@ -21,12 +26,7 @@ public class CommentsServiceImpl implements CommentsService {
     private final UsersRepository usersRepository;
     private final CommentConverter commentConverter;
 
-    public CommentsServiceImpl(CommentsRepository commentsRepository, TasksRepository tasksRepository, UsersRepository usersRepository, CommentConverter commentConverter) {
-        this.commentsRepository = commentsRepository;
-        this.tasksRepository = tasksRepository;
-        this.usersRepository = usersRepository;
-        this.commentConverter = commentConverter;
-    }
+
 
     @Override
     public List<CommentResponse> findAll(int page) {
@@ -39,7 +39,7 @@ public class CommentsServiceImpl implements CommentsService {
     }
 
     @Override
-    public CommentResponse publishComment(CommentRequest commentRequest, int taskId) {
+    public CommentResponse publishComment(CommentRequest commentRequest, int taskId,String userEmail) {
         Comment comment = commentConverter.convertToComment(commentRequest);
         comment.setAuthor(usersRepository.findByEmail("dsds").orElseThrow());//todo principal to user and exception
         comment.setCreatedAt(new Date());
@@ -48,17 +48,21 @@ public class CommentsServiceImpl implements CommentsService {
     }
 
     @Override
-    @PreFilter("@userServiceImpl.isEnoughRules(commentId,principal.username)")
-    public void update(int commentId, CommentRequest commentRequest) {
-        Comment updatedComment = commentsRepository.findById(commentId).orElseThrow();//todo exception
-        Comment changes = commentConverter.convertToComment(commentRequest);
-        updatedComment.setText(changes.getText());
+    public void update(int commentId, CommentRequest commentRequest,String userEmail) {
+
+        if(isEnoughRules(commentId,userEmail)){
+            Comment updatedComment = commentsRepository.findById(commentId).orElseThrow();//todo exception
+            Comment changes = commentConverter.convertToComment(commentRequest);
+            updatedComment.setText(changes.getText());
+        } else throw new RuntimeException();
     }
 
     @Override
-    @PreFilter("@userServiceImpl.isEnoughRules(commentId,principal.username)")
-    public void delete(int commentId) {
-        commentsRepository.deleteById(commentId);
+    public void delete(int commentId,String userEmail) {
+        if(isEnoughRules(commentId,userEmail)){
+            commentsRepository.deleteById(commentId);
+        } else throw new RuntimeException();
+
     }
 
     @Override
